requirements.txt
absl-py==2.0.0
apturl==0.5.2
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
deepface==0.0.79
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
djangorestframework==3.14.0
dlib==19.24.2
duplicity==0.8.21
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
future==0.18.2
gast==0.4.0
gdown==4.7.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
httplib2==0.20.2
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
imutils==0.5.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mediapipe==0.10.7
ml-dtypes==0.2.0
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
msgpack==1.0.7
mtcnn==0.1.1
mysqlclient==2.1.1
netifaces==0.11.0
nltk==3.8.1
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.3
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.8.0
proglog==0.1.10
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydub==0.25.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
QtPy==2.0.0
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==3.3.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.0
sqlparse==0.4.4
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-cpu==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.6
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==3.17.0
